# api/docs/swagger.yaml
definitions:
...
  company_service.AddResourceRequest:
    properties:
      company_id:
        type: string
      credentials:
        $ref: '#/definitions/company_service.Resource_Credentials'
      environment_id:
        type: string
      is_configured:
        type: boolean
      project_id:
        type: string
      resource_type:
        type: integer
      service_type:
        type: integer
      title:
        type: string
      user_id:
        type: string
    type: object
  company_service.AddResourceResponse:
    properties:
      id:
        type: string
    type: object
...
  company_service.ConfigureResourceRequest:
    properties:
      company_id:
        type: string
      credentials:
        $ref: '#/definitions/company_service.Resource_Credentials'
      environment_id:
        type: string
      is_configured:
        type: boolean
      project_id:
        type: string
      resource_id:
        type: string
      resource_type:
        type: integer
      service_type:
        type: integer
      title:
        type: string
      user_id:
        type: string
    type: object
  company_service.ConfigureResourceResponse:
    properties:
      id:
        type: string
    type: object
...
  company_service.CreateResourceReq:
    properties:
      company_id:
        type: string
      project_id:
        type: string
      resource:
        $ref: '#/definitions/company_service.Resource'
      user_id:
        type: string
    type: object
  company_service.CreateResourceRes:
    properties:
      company_id:
        type: string
      project_id:
        type: string
      resource_type:
        type: integer
      service_type:
        type: integer
    type: object
  company_service.EmptyProto:
    type: object
...
  company_service.GetReourceListResponse:
    properties:
      count:
        type: integer
      resources:
        items:
          $ref: '#/definitions/company_service.NewResource'
        type: array
    type: object
  company_service.NewResource:
    properties:
      id:
        type: string
      project_id:
        type: string
      title:
        type: string
    type: object
...
  company_service.ReconnectResourceRequest:
    properties:
      id:
        type: string
    type: object
  company_service.RemoveResourceRequest:
    properties:
      id:
        type: string
    type: object
  company_service.Resource:
    properties:
      credentials:
        $ref: '#/definitions/company_service.Resource_Credentials'
      id:
        type: string
      resource_type:
        type: integer
      service_type:
        type: integer
      title:
        type: string
    type: object
  company_service.Resource_Credentials:
    properties:
      database:
        type: string
      host:
        type: string
      password:
        type: string
      port:
        type: string
      username:
        type: string
    type: object
  company_service.ResourceWithoutPassword:
    properties:
      credentials:
        $ref: '#/definitions/company_service.ResourceWithoutPassword_Credentials'
      environments:
        items:
          $ref: '#/definitions/company_service.ResourceWithoutPassword_Environment'
        type: array
      id:
        type: string
      project_id:
        type: string
      resource_type:
        type: integer
      title:
        type: string
    type: object
  company_service.ResourceWithoutPassword_Credentials:
    properties:
      database:
        type: string
      host:
        type: string
      port:
        type: string
      username:
        type: string
    type: object
  company_service.ResourceWithoutPassword_Environment:
    properties:
      description:
        type: string
      display_color:
        type: string
      id:
        type: string
      is_configured:
        type: boolean
      name:
        type: string
      resource_environment_id:
        type: string
    type: object
  company_service.UpdateResourceRequest:
    properties:
      credentials:
        $ref: '#/definitions/company_service.Resource_Credentials'
      environment_id:
        type: string
      id:
        type: string
      is_configured:
        type: boolean
      project_id:
        type: string
      resource_type:
        type: integer
      title:
        type: string
    type: object
...
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a api gateway
  termsOfService: https://udevs.io
paths:
...
  /v1/company/project/configure-resource:
    post:
      consumes:
      - application/json
      description: Configure ProjectResource
      operationId: configure_project_resource
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: ProjectResourceConfigureRequest
        in: body
        name: ProjectResource
        required: true
        schema:
          $ref: '#/definitions/company_service.ConfigureResourceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ProjectResource data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/company_service.ConfigureResourceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Configure ProjectResource
      tags:
      - Company Resource
  /v1/company/project/create-resource:
    post:
      consumes:
      - application/json
      description: Create ProjectResource
      operationId: create_project_resource
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: ProjectResourceCreateRequest
        in: body
        name: ProjectResource
        required: true
        schema:
          $ref: '#/definitions/company_service.CreateResourceReq'
      produces:
      - application/json
      responses:
        "201":
          description: ProjectResource data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/company_service.CreateResourceRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create ProjectResource
      tags:
      - Company Resource
  /v1/company/project/resource:
    delete:
      consumes:
      - application/json
      description: Remove ProjectResource
      operationId: remove_project_resource
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: ProjectResourceRemoveRequest
        in: body
        name: ProjectResource
        required: true
        schema:
          $ref: '#/definitions/company_service.RemoveResourceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ProjectResource data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/company_service.EmptyProto'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Remove ProjectResource
      tags:
      - Company Resource
    get:
      consumes:
      - application/json
      description: Get all companies
      operationId: get_resource_list
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      - in: query
        name: project_id
        type: string
      - in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Resource data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/company_service.GetReourceListResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get all companies
      tags:
      - Company Resource
    post:
      consumes:
      - application/json
      description: Add ProjectResource
      operationId: add_project_resource
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: ProjectResourceAddRequest
        in: body
        name: ProjectResource
        required: true
        schema:
          $ref: '#/definitions/company_service.AddResourceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ProjectResource data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/company_service.AddResourceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Add ProjectResource
      tags:
      - Company Resource
  /v1/company/project/resource-environment/{resource_id}:
    get:
      consumes:
      - application/json
      description: Get Resource Environment by id
      operationId: get_resource_environment_id
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: resource_id
        in: path
        name: resource_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Resource data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/company_service.ResourceWithoutPassword'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Resource Environment by id
      tags:
      - Company Resource
  /v1/company/project/resource/{resource_id}:
    get:
      consumes:
      - application/json
      description: Get Resource by id
      operationId: get_resource_id
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: resource_id
        in: path
        name: resource_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Resource data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/company_service.ResourceWithoutPassword'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Resource by id
      tags:
      - Company Resource
    put:
      consumes:
      - application/json
      description: Update Resource by id
      operationId: put_resource_id
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: UpdateResourceRequestBody
        in: body
        name: UpdateResourceRequestBody
        required: true
        schema:
          $ref: '#/definitions/company_service.UpdateResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Resource data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/company_service.ResourceWithoutPassword'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update Resource by id
      tags:
      - Company Resource
  /v1/company/project/resource/reconnect:
    post:
      consumes:
      - application/json
      description: Reconnect ProjectResource
      operationId: reconnect_project_resource
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: ProjectResourceReconnectRequest
        in: body
        name: ProjectResource
        required: true
        schema:
          $ref: '#/definitions/company_service.ReconnectResourceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ProjectResource data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/company_service.EmptyProto'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Reconnect ProjectResource
      tags:
      - Company Resource
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"